Split candidate list repository
