+++ src/candidate_lists/candidate_pages/edit_position.rs
@@ -78,7 +78,7 @@ pub async fn update_candidate_position(
 
             match position_form.action {
                 CandidatePositionAction::Remove => {
-                    candidate_lists::repository::remove_candidate(
+                    candidate_lists::remove_candidate(
                         &mut conn,
                         candidate.list_id,
                         candidate.person.id,
@@ -88,7 +88,7 @@ pub async fn update_candidate_position(
                 CandidatePositionAction::Move => {
                     let mut full_list = full_list;
                     full_list.update_position(candidate.person.id, position_form.position);
-                    candidate_lists::repository::update_candidate_list_order(
+                    candidate_lists::update_candidate_list_order(
                         &mut conn,
                         candidate.list_id,
                         &full_list.get_ids(),
@@ -138,16 +138,16 @@ mod tests {
         let person = sample_person(PersonId::new());
 
         let mut conn = pool.acquire().await?;
-        candidate_lists::repository::create_candidate_list(&mut conn, &list).await?;
-        persons::repository::create_person(&mut conn, &person).await?;
-        candidate_lists::repository::update_candidate_list_order(&mut conn, list_id, &[person.id])
+        candidate_lists::create_candidate_list(&mut conn, &list).await?;
+        persons::create_person(&mut conn, &person).await?;
+        candidate_lists::update_candidate_list_order(&mut conn, list_id, &[person.id])
             .await?;
 
-        let full_list = candidate_lists::repository::get_full_candidate_list(&mut conn, list_id)
+        let full_list = candidate_lists::get_full_candidate_list(&mut conn, list_id)
             .await?
             .expect("candidate list");
         let candidate =
-            candidate_lists::repository::get_candidate(&mut conn, list_id, person.id).await?;
+            candidate_lists::get_candidate(&mut conn, list_id, person.id).await?;
 
         let response = edit_candidate_position(
             EditCandidatePositionPath {
@@ -179,21 +179,21 @@ mod tests {
         let person_b = sample_person_with_last_name(PersonId::new(), "Bakker");
 
         let mut conn = pool.acquire().await?;
-        candidate_lists::repository::create_candidate_list(&mut conn, &list).await?;
-        persons::repository::create_person(&mut conn, &person_a).await?;
-        persons::repository::create_person(&mut conn, &person_b).await?;
-        candidate_lists::repository::update_candidate_list_order(
+        candidate_lists::create_candidate_list(&mut conn, &list).await?;
+        persons::create_person(&mut conn, &person_a).await?;
+        persons::create_person(&mut conn, &person_b).await?;
+        candidate_lists::update_candidate_list_order(
             &mut conn,
             list_id,
             &[person_a.id, person_b.id],
         )
         .await?;
 
-        let full_list = candidate_lists::repository::get_full_candidate_list(&mut conn, list_id)
+        let full_list = candidate_lists::get_full_candidate_list(&mut conn, list_id)
             .await?
             .expect("candidate list");
         let candidate =
-            candidate_lists::repository::get_candidate(&mut conn, list_id, person_a.id).await?;
+            candidate_lists::get_candidate(&mut conn, list_id, person_a.id).await?;
 
         let csrf_tokens = CsrfTokens::default();
         let csrf_token = csrf_tokens.issue().value;
@@ -218,7 +218,7 @@ mod tests {
         assert_eq!(response.status(), StatusCode::SEE_OTHER);
 
         let mut conn = pool.acquire().await?;
-        let full_list = candidate_lists::repository::get_full_candidate_list(&mut conn, list_id)
+        let full_list = candidate_lists::get_full_candidate_list(&mut conn, list_id)
             .await?
             .expect("candidate list");
         assert_eq!(full_list.candidates.len(), 2);
@@ -236,21 +236,21 @@ mod tests {
         let person_b = sample_person_with_last_name(PersonId::new(), "Bakker");
 
         let mut conn = pool.acquire().await?;
-        candidate_lists::repository::create_candidate_list(&mut conn, &list).await?;
-        persons::repository::create_person(&mut conn, &person_a).await?;
-        persons::repository::create_person(&mut conn, &person_b).await?;
-        candidate_lists::repository::update_candidate_list_order(
+        candidate_lists::create_candidate_list(&mut conn, &list).await?;
+        persons::create_person(&mut conn, &person_a).await?;
+        persons::create_person(&mut conn, &person_b).await?;
+        candidate_lists::update_candidate_list_order(
             &mut conn,
             list_id,
             &[person_a.id, person_b.id],
         )
         .await?;
 
-        let full_list = candidate_lists::repository::get_full_candidate_list(&mut conn, list_id)
+        let full_list = candidate_lists::get_full_candidate_list(&mut conn, list_id)
             .await?
             .expect("candidate list");
         let candidate =
-            candidate_lists::repository::get_candidate(&mut conn, list_id, person_a.id).await?;
+            candidate_lists::get_candidate(&mut conn, list_id, person_a.id).await?;
 
         let csrf_tokens = CsrfTokens::default();
         let csrf_token = csrf_tokens.issue().value;
@@ -275,7 +275,7 @@ mod tests {
         assert_eq!(response.status(), StatusCode::SEE_OTHER);
 
         let mut conn = pool.acquire().await?;
-        let full_list = candidate_lists::repository::get_full_candidate_list(&mut conn, list_id)
+        let full_list = candidate_lists::get_full_candidate_list(&mut conn, list_id)
             .await?
             .expect("candidate list");
         assert_eq!(full_list.candidates.len(), 1);
@@ -294,21 +294,21 @@ mod tests {
         let person_b = sample_person_with_last_name(PersonId::new(), "Bakker");
 
         let mut conn = pool.acquire().await?;
-        candidate_lists::repository::create_candidate_list(&mut conn, &list).await?;
-        persons::repository::create_person(&mut conn, &person_a).await?;
-        persons::repository::create_person(&mut conn, &person_b).await?;
-        candidate_lists::repository::update_candidate_list_order(
+        candidate_lists::create_candidate_list(&mut conn, &list).await?;
+        persons::create_person(&mut conn, &person_a).await?;
+        persons::create_person(&mut conn, &person_b).await?;
+        candidate_lists::update_candidate_list_order(
             &mut conn,
             list_id,
             &[person_a.id, person_b.id],
         )
         .await?;
 
-        let full_list = candidate_lists::repository::get_full_candidate_list(&mut conn, list_id)
+        let full_list = candidate_lists::get_full_candidate_list(&mut conn, list_id)
             .await?
             .expect("candidate list");
         let candidate =
-            candidate_lists::repository::get_candidate(&mut conn, list_id, person_a.id).await?;
+            candidate_lists::get_candidate(&mut conn, list_id, person_a.id).await?;
 
         let csrf_tokens = CsrfTokens::default();
         let other_tokens = CsrfTokens::default();
@@ -336,7 +336,7 @@ mod tests {
         assert!(body.contains("The CSRF token is invalid."));
 
         let mut conn = pool.acquire().await?;
-        let full_list = candidate_lists::repository::get_full_candidate_list(&mut conn, list_id)
+        let full_list = candidate_lists::get_full_candidate_list(&mut conn, list_id)
             .await?
             .expect("candidate list");
         assert_eq!(full_list.candidates.len(), 2);
