+++ src/candidate_lists/mod.rs
@@ -2,9 +2,11 @@ mod candidate_pages;
 mod extractors;
 mod pages;
 mod structs;
+mod repository;
+mod candidate_repository;
 
-pub mod repository;
-
-pub use candidate_pages::candidate_router;
 pub use pages::router;
+pub use candidate_pages::candidate_router;
 pub use structs::*;
+pub use repository::*;
+pub use candidate_repository::*;
