+++ src/candidate_lists/pages/update.rs
@@ -30,8 +30,8 @@ pub async fn edit_candidate_list(
     DbConnection(mut conn): DbConnection,
 ) -> Result<Response, AppError> {
     let candidate_lists =
-        candidate_lists::repository::list_candidate_list_with_count(&mut conn).await?;
-    let total_persons = persons::repository::count_persons(&mut conn).await?;
+        candidate_lists::list_candidate_list_with_count(&mut conn).await?;
+    let total_persons = persons::count_persons(&mut conn).await?;
 
     Ok(HtmlTemplate(
         CandidateListUpdateTemplate {
@@ -57,8 +57,8 @@ pub async fn update_candidate_list(
     form: Form<CandidateListForm>,
 ) -> Result<Response, AppError> {
     let candidate_lists =
-        candidate_lists::repository::list_candidate_list_with_count(&mut conn).await?;
-    let total_persons = persons::repository::count_persons(&mut conn).await?;
+        candidate_lists::list_candidate_list_with_count(&mut conn).await?;
+    let total_persons = persons::count_persons(&mut conn).await?;
 
     match form.validate_update(&candidate_list, &csrf_tokens) {
         Err(form_data) => Ok(HtmlTemplate(
@@ -73,7 +73,7 @@ pub async fn update_candidate_list(
         .into_response()),
         Ok(candidate_list) => {
             let candidate_list =
-                candidate_lists::repository::update_candidate_list(&mut conn, candidate_list)
+                candidate_lists::update_candidate_list(&mut conn, candidate_list)
                     .await?;
             Ok(Redirect::to(&candidate_list.view_path()).into_response())
         }
@@ -99,7 +99,7 @@ mod tests {
         let candidate_list = sample_candidate_list(CandidateListId::new());
 
         let mut conn = pool.acquire().await?;
-        candidate_lists::repository::create_candidate_list(&mut conn, &candidate_list).await?;
+        candidate_lists::create_candidate_list(&mut conn, &candidate_list).await?;
 
         let response = edit_candidate_list(
             CandidateListsEditPath {
@@ -135,7 +135,7 @@ mod tests {
             created_at: creation_date,
             updated_at: creation_date,
         };
-        candidate_lists::repository::create_candidate_list(&mut conn, &candidate_list).await?;
+        candidate_lists::create_candidate_list(&mut conn, &candidate_list).await?;
 
         let form = CandidateListForm {
             electoral_districts: vec![ElectoralDistrict::DR],
@@ -165,7 +165,7 @@ mod tests {
 
         // verify updated candidate list object in database
         let mut conn = pool.acquire().await?;
-        let lists = candidate_lists::repository::list_candidate_list_with_count(&mut conn).await?;
+        let lists = candidate_lists::list_candidate_list_with_count(&mut conn).await?;
         assert_eq!(lists.len(), 1);
 
         let updated_list = &lists[0].list;
@@ -198,7 +198,7 @@ mod tests {
             created_at: creation_date,
             updated_at: creation_date,
         };
-        candidate_lists::repository::create_candidate_list(&mut conn, &candidate_list).await?;
+        candidate_lists::create_candidate_list(&mut conn, &candidate_list).await?;
 
         let form = CandidateListForm {
             electoral_districts: vec![ElectoralDistrict::DR],
@@ -222,7 +222,7 @@ mod tests {
         assert!(body.contains("Edit candidate list"));
 
         let mut conn = pool.acquire().await?;
-        let lists = candidate_lists::repository::list_candidate_list_with_count(&mut conn).await?;
+        let lists = candidate_lists::list_candidate_list_with_count(&mut conn).await?;
         assert_eq!(lists.len(), 1);
 
         let updated_list = &lists[0].list;
